Separate action and reducer
